// state/Branches/branches.action-types.ts
export enum ActionTypes {
    BRANCHES_UPLOAD_REQUEST = 'BRANCHES_UPLOAD_REQUEST',
    BRANCHES_UPLOAD_SUCCESS = 'BRANCHES_UPLOAD_SUCCESS',
    BRANCHES_UPLOAD_FAIL = 'BRANCHES_UPLOAD_FAIL',
  
    BRANCHES_FETCH_REQUEST = 'BRANCHES_FETCH_REQUEST',
    BRANCHES_FETCH_SUCCESS = 'BRANCHES_FETCH_SUCCESS',
    BRANCHES_FETCH_FAIL = 'BRANCHES_FETCH_FAIL',
  
    BRANCHES_DELETE_REQUEST = 'BRANCHES_DELETE_REQUEST',
    BRANCHES_DELETE_SUCCESS = 'BRANCHES_DELETE_SUCCESS',
    BRANCHES_DELETE_FAIL = 'BRANCHES_DELETE_FAIL',
  
    BRANCHES_EDIT_REQUEST = 'BRANCHES_EDIT_REQUEST',
    BRANCHES_EDIT_SUCCESS = 'BRANCHES_EDIT_SUCCESS',
    BRANCHES_EDIT_FAIL = 'BRANCHES_EDIT_FAIL',
  
    BRANCHES_RESET = 'BRANCHES_RESET',
  }
  